Refactor query processing interfaces
