+++ core/src/main/java/org/polypheny/db/transaction/Transaction.java
@@ -19,8 +19,6 @@ package org.polypheny.db.transaction;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
-import org.polypheny.db.QueryProcessor;
-import org.polypheny.db.SqlProcessor;
 import org.polypheny.db.adapter.DataContext;
 import org.polypheny.db.adapter.Store;
 import org.polypheny.db.adapter.java.JavaTypeFactory;
@@ -29,6 +27,8 @@ import org.polypheny.db.information.InformationDuration;
 import org.polypheny.db.information.InformationManager;
 import org.polypheny.db.jdbc.Context;
 import org.polypheny.db.prepare.PolyphenyDbCatalogReader;
+import org.polypheny.db.processing.QueryProcessor;
+import org.polypheny.db.processing.SqlProcessor;
 import org.polypheny.db.routing.Router;
 import org.polypheny.db.schema.PolyphenyDbSchema;
 import org.polypheny.db.sql.parser.SqlParser.SqlParserConfig;
